requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
diskcache==5.6.3
distro==1.8.0
docformatter==1.7.5
docker==6.1.3
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.13.1
flake8==6.1.0
FLAML==2.1.1
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
litellm==0.13.1
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pyautogen==0.1.14
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
termcolor==2.3.0
tiktoken==0.5.1
tokenizers==0.14.1
tomli==2.0.1
tomli_w==1.0.0
tqdm==4.66.1
typing_extensions==4.8.0
untokenize==0.1.1
urllib3==2.0.7
uvicorn==0.24.0
websocket-client==1.6.4
yarl==1.9.2
zipp==3.17.0
